Move remaining 'library' folder items to app/
