+++ jsconfig.json
@@ -2,5 +2,6 @@
 	"compilerOptions": {
 		"checkJs": true,
 		"target": "es2024", // es6, es2024
+		"module": "es2022"
 	}
 }
